refactor: components and folder structure
