--- Packages/com.zelcam4.ml-agents/Runtime/Actuators.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5a8efb61608d6b64b843426e2f52cc6f
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
