--- Packages/com.zelcam4.ml-agents/Runtime/Actuators/VectorActuator.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 6afefba1b981a714db0bb6b41941d57e
-timeCreated: 1593023833
