--- Packages/com.zelcam4.ml-agents/Runtime/Agent.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3d9e7a39f4db33a4dabda52ca8ec2ddd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
