+++ Packages/com.zelcam4.ml-agents/Runtime/Autoring/AgentAutoring.cs
@@ -1,5 +1,5 @@
 using Unity.Entities;
-using Zelcam4.MLAgents.Actuators;
+
 using UnityEngine;
 
 namespace Zelcam4.MLAgents
