+++ Packages/com.zelcam4.ml-agents/Runtime/Communicator/CommunicatorManager.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using Unity.Collections;
 using Unity.Entities;
 using Zelcam4.MLAgents;
-using Zelcam4.MLAgents.Actuators;
+
 using UnityEngine;
 using Zelcam4.MLAgents.CommunicatorObjects;
