+++ Packages/com.zelcam4.ml-agents/Runtime/Communicator/ICommunicator.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 using Unity.Entities;
-using Zelcam4.MLAgents.Actuators;
+
 
 
 namespace Zelcam4.MLAgents
