+++ Packages/com.zelcam4.ml-agents/Runtime/Communicator/RpcCommunicator.cs
@@ -15,7 +15,7 @@ using Zelcam4.MLAgents.CommunicatorObjects;
 using Google.Protobuf;
 using Unity.Entities;
 using Zelcam4.MLAgents;
-using Zelcam4.MLAgents.Actuators;
+
