+++ Packages/com.zelcam4.ml-agents/Runtime/Data/Agent.cs
@@ -1,0 +1,29 @@
+using Unity.Entities;
+
+namespace Zelcam4.MLAgents
+{
+    public struct AgentEcs : IComponentData
+    {
+        public float Reward;
+        public float GroupReward;
+        public float CumulativeReward;
+        public bool RequestAction;   
+        public bool RequestDecision; 
+        public int MaxStep;
+        public int StepCount;
+        public int CompletedEpisodes;
+        public int EpisodeId;
+        public bool Initialized;
+
+        public int GroupId;
+        public bool IsEnabled;      
+
+        //Adding From AgentInfo
+        public bool MaxStepReached;
+        public bool Done;
+        public bool StartingEpisode;
+
+        public Entity Target;
+        public Entity GoundRender;
+    }
+}
