+++ Packages/com.zelcam4.ml-agents/Runtime/Data/Brain.cs
@@ -2,7 +2,7 @@ using System.Linq;
 using Unity.Collections;
 using Unity.Entities;
 
-using Zelcam4.MLAgents.Actuators;
+
 
 namespace Zelcam4.MLAgents
 {
