--- Packages/com.zelcam4.ml-agents/Runtime/Data/Communicator.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: eb00547327d2b364b9343b170a79dff6
