+++ Packages/com.zelcam4.ml-agents/Runtime/Data/Observation.cs
@@ -3,6 +3,15 @@ using Unity.Mathematics;
 
 namespace Zelcam4.MLAgents
 {
+    public struct ObservationValue : IBufferElementData
+    {
+        public float Value;
+        
+        // Convenience
+        public static implicit operator float(ObservationValue e) { return e.Value; }
+        public static implicit operator ObservationValue(float f) { return new ObservationValue { Value = f }; }
+    }
+    
     public enum ObservationSourceType
     {
         PositionX, PositionY, PositionZ,
