--- Packages/com.zelcam4.ml-agents/Runtime/Inference.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 6e4fd6cb41632a9408af19fc7a2b76bb
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
