--- Packages/com.zelcam4.ml-agents/Runtime/Inference/Utils.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9cf87f4152d58374a88584e01979150d
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
