--- Packages/com.zelcam4.ml-agents/Runtime/Inference/Utils/Multinomial.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f4c7b45d2a999e94798665162ecf2dca
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
