--- Packages/com.zelcam4.ml-agents/Runtime/Inference/Utils/RandomNormal.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a1dc601de425da945b9654cb36d44f22
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
