--- Packages/com.zelcam4.ml-agents/Runtime/MLAgentsSettings.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5d23bc771747b2f4cb1ff0e2a604c04b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
