--- Packages/com.zelcam4.ml-agents/Runtime/MLAgentsSettingsManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9fc43fb092f891e49aa336b5492c5dc0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
