--- Packages/com.zelcam4.ml-agents/Runtime/Managers.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0e52bad24d229b4449f2833e65301314
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
