+++ Packages/com.zelcam4.ml-agents/Runtime/Systems/AgentActionSystem.cs
@@ -2,7 +2,7 @@ using Unity.Burst;
 using Unity.Entities;
 using Unity.Mathematics;
 using Zelcam4.MLAgents;
-using Zelcam4.MLAgents.Actuators;
+
 using Unity.Transforms;
 using UnityEngine;
 using static Unity.Entities.SystemAPI;
