+++ Packages/com.zelcam4.ml-agents/Runtime/Systems/DecideActionSystem.cs
@@ -3,7 +3,7 @@ using Unity.Burst;
 using Unity.Collections;
 using Unity.Entities;
 using Zelcam4.MLAgents;
-using Zelcam4.MLAgents.Actuators;
+
 using UnityEngine;
