--- Packages/com.zelcam4.ml-agents/Runtime/UnityAgentsException.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 0618d0b5398ba5040b087551a09470e4
-timeCreated: 1504131359
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
